fix(desktop): retire single-member built-in Fizz team
A built-in team with one member serves no purpose, and built-ins are
deletion-protected so users couldn't remove it themselves. On load,
stored copies still matching the seed are purged; customized copies
demote to user-owned teams via the existing retirement path. The
builtin:fizz agent itself is unaffected.
